added neovim configs

+++ install_dotfiles.sh
@@ -11,9 +11,17 @@ fi
 WORKDIR=$PWD
 cd $DIR
 for f in `command ls -ap | grep -v /`; do
-    echo "Creating a symbolic link of $f in $HOME"
-    ln -s "$DIR/$f" "$HOME/$f"
+    if [ -f "$HOME/$f" ]; then
+        echo "$HOME/$f: Symbolic link already exists."
+    else
+        echo "Creating a symbolic link of $f in $HOME"
+        ln -s "$DIR/$f" "$HOME/$f"
+    fi
 done
+if [ -d ./nvim ]; then
+    mkdir -p "$HOME/.config/nvim/session"
+    cp -rs "$DIR/nvim" "$HOME/.config"
+fi
 
 cd $WORKDIR
